% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pallette.R
\name{find_pallette}
\alias{find_pallette}
\title{Find the pallette of your dreams}
\usage{
find_pallette(min_colours = 0, colourblind_friendly = FALSE, type = NULL)
}
\arguments{
\item{min_colours}{Minimum number of colours that the pallette must support}

\item{colourblind_friendly}{Does the pallette need to be colourblind friendly}

\item{type}{What type? Allowed values are "categorical", "diverging_discrete" or "sequential_discrete"}
}
\value{
A tibble of metadata describing the pallettes found
}
\description{
Find the pallette of your dreams
}
\examples{
find_pallette()
find_pallette(min_colors=12)
find_pallette(type="categorical")
}
