% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colours.R
\name{get_colours}
\alias{get_colours}
\title{Get the colours from a specific pallette}
\usage{
get_colours(short_name, num_colours = 0)
}
\arguments{
\item{short_name}{The pallette short name}

\item{num_colours}{Number of colours to return}
}
\value{
A vector of colours
}
\description{
Get the colours from a specific pallette
}
\examples{
get_colours("haicdf")
get_colours("ptccc", num_colours=4)
}
